Movi a notificação para um modelo a parte, para facilitar o entendimento do código
