+++ src/main/java/ds/desktop/notify/DesktopNotify.java
@@ -43,7 +43,7 @@ public class DesktopNotify {
     public static final int LEFT_TO_RIGHT = 0;
     public static final int RIGHT_TO_LEFT = 1;
 
-    private static NotifyTheme defTheme = NotifyTheme.Dark;
+    public static NotifyTheme defTheme = NotifyTheme.Dark;
 
     /**
      * Sets the theme to use by default when creating notifications.
@@ -265,445 +265,4 @@ public class DesktopNotify {
         pane.show();
     }
 
-    String title;
-    String message;
-    Image icon;
-    int type;
-    int orientation;
-
-    /**
-     * A <code>NotifyTheme</code> object with the color and Font parameters to
-     * be used when this notification is being painted.
-     */
-    NotifyTheme theme;
-
-    /**
-     * A String array with the lines of the title.
-     * It is refreshed automatically when the title is set.
-     */
-    String[] tlts = new String[0];
-    /**
-     * A String array with the lines of the message.
-     * It is refreshed automatically when the message is set.
-     */
-    String[] msgs = new String[0];
-
-    int w = 0;
-    int h = 0;
-    /**
-     * -- GETTER --
-     *
-     * @return <code>true</code> if this notification is currently being shown.
-     */
-    @Getter
-    boolean visible = false;
-    boolean markedForHide = false;
-
-    int highl = 0;
-
-    long popupStart = 0L;
-    long timeOut = 8000L;
-
-    /**
-     * An action to perform when the notification is clicked.
-     * -- SETTER --
-     *  Allows to bind an action to this notification. It will be executed when
-     *  the notification gets clicked. Notifications with an action set will wait
-     *  for the user to click them.
-     *
-     *
-     * -- GETTER --
-     *
-     @param action he ActionListener to bind to this notification.
-      * @return the ActionListener binded to this notification.
-
-     */
-    @Getter
-    @Setter
-    ActionListener action;
-
-    /**
-     * A protected constructor for a DesktopNotify object, called internally.
-     * You can use any of the <code>showDesktopMessage()</code> static methods
-     * or a {@link NotificationBuilder} in order to create and display a
-     * notification instead.
-     *
-     * @param title       The notification title. Can be {@code null} when the
-     *                    notification isn't meant to have one.
-     * @param message     The notification message. Can be {@code null} when
-     *                    the notification isn't meant to have one.
-     * @param orientation The text orientation, either {@link #LEFT_TO_RIGHT} or
-     *                    {@link #RIGHT_TO_LEFT}.
-     * @param type        The notification type. Check the type contans in the
-     *                    class for the ones you can use.
-     * @param icon        The icon to display. Can be {@code null} when the
-     *                    notification isn't meant to have one.
-     */
-    protected DesktopNotify(String title, String message, int type,
-                            int orientation, Image icon) {
-        this.title = (title == null ? "" : title);
-        this.message = (message == null ? "" : message);
-        this.type = type;
-        this.orientation = orientation;
-        this.icon = icon;
-        this.theme = DesktopNotify.defTheme;
-    }
-
-    /**
-     * Allows to set a <code>NotifyTheme</code> for this notification. One of
-     * these defines stuff as the colors for text and background, the fonts and
-     * even images for the default icons.
-     *
-     * @param theme The <code>NotifyTheme</code> to use for this notification.
-     */
-    public void setTheme(NotifyTheme theme) {
-        if (theme != null) this.theme = theme;
-    }
-
-    /**
-     * Allows to set a timeout for this notification, in milliseconds.
-     * After the time is 'out', and if there isn't an action set, the
-     * notification will just fade away.
-     *
-     * @param millis the timeout.
-     */
-    public void setTimeout(long millis) {
-        timeOut = millis < 0 ? 0 : millis;
-    }
-
-    protected long expTime() {
-        return timeOut == 0 ? Long.MAX_VALUE : popupStart + timeOut;
-    }
-
-    /**
-     * Polls the Driver in order to show this notification, so you don't have to
-     * do it yourself.
-     */
-    public void show() {
-        markedForHide = false;
-        DesktopNotifyDriver.postPane(this);
-    }
-
-    /**
-     * Hides this notification immediately if it is already being displayed.
-     * The action itself is handled by the Driver, this method only marks this
-     * notification for hiding.
-     */
-    public void hide() {
-        markedForHide = true;
-    }
-
-    protected void setWidth(int w) {
-        this.w = w;
-    }
-
-    /**
-     * Used by he driver to mark this notification as visible or not.
-     *
-     * @param visible {@code true} if the notification is currently visible.
-     */
-    protected void setVisible(boolean visible) {
-        this.visible = visible;
-    }
-
-    /**
-     * The painting routine for the notification goes here.
-     *
-     * @param x     Where to start painting (X)
-     * @param y     Where to start painting (Y)
-     * @param hover A boolean that is <code>true</code> if the user is hovering
-     *              this notification with the mouse pointer.
-     * @param rd    The <code>Graphics2D</code> object to use for painting.
-     * @param l     The current time.
-     */
-    public void render(int x, int y, boolean hover, Graphics2D rd, long l) {
-        long i = l - popupStart;
-        if (i > 500) i = expTime() - l;
-        if (i < 0) i = 0;
-        if (i > 500) i = -1;
-        rd.setRenderingHint(RenderingHints.KEY_TEXT_ANTIALIASING,
-                RenderingHints.VALUE_TEXT_ANTIALIAS_ON);
-        AffineTransform trans = rd.getTransform();
-        rd.translate(x, y);
-        if (i != -1) {
-            double d = i / 500.0;
-            rd.translate(w / 2 - ((w / 2) * d), h / 2 - ((h / 2) * d));
-            rd.scale(d, d);
-            rd.setComposite(AlphaComposite
-                    .getInstance(AlphaComposite.SRC_OVER, (float) d));
-        }
-        rd.setPaint(new GradientPaint(0, (title.isEmpty() ? 0 : 25),
-                /*hover? new Color(50,57,65):*/theme.bgGrad[0],
-                0, h, theme.bgGrad[highl + 1], false));
-        if (hover && highl < 20) highl++;
-        if (!hover && highl > 0) highl--;
-        rd.fillRoundRect(0, 0, w, h, 20, 20);
-        rd.setPaint(null);
-        rd.setColor(theme.borderColor);
-        rd.drawRoundRect(0, 0, w - 1, h - 1, 20,20);
-        if (i == -1) {
-            int titleH = getLineHeight(theme.titleFont);
-            int textH = getLineHeight(theme.contentFont);
-            if (!title.isEmpty()) {
-                rd.setColor(theme.titleColor);
-                rd.setFont(theme.titleFont);
-                int tX = 5 + ((icon == null && type == 0) ? 0 : 38);
-                for (int j = 0; j < tlts.length; j++) {
-                    if (orientation == RIGHT_TO_LEFT) {
-                        FontMetrics ftm = DesktopNotifyDriver.getFontMetrics(theme.titleFont);
-                        tX = w - 4 - ((icon == null && type == 0) ? 0 : 38) - ftm.stringWidth(tlts[j]);
-                    }
-                    rd.drawString(tlts[j], tX, 20 + (titleH * j));
-                }
-            }
-            if (!message.isEmpty()) {
-                rd.setColor(theme.contentColor);
-                rd.setFont(theme.contentFont);
-                int tX = 6 + ((icon == null && type == 0) ? 0 : 38);
-                for (int j = 0; j < msgs.length; j++) {
-                    if (orientation == RIGHT_TO_LEFT) {
-                        FontMetrics ftm = DesktopNotifyDriver.getFontMetrics(theme.contentFont);
-                        tX = w - 5 - ((icon == null && type == 0) ? 0 : 38) - ftm.stringWidth(msgs[j]);
-                    }
-                    rd.drawString(msgs[j], tX, 20 + (titleH * tlts.length) + (textH * j));
-                }
-            }
-        }
-//        if(expTime()==Long.MAX_VALUE){
-//            rd.setFont(theme.titleFont);
-//            rd.setColor(new Color(120,120,120));
-//            rd.drawString("X", w-16, 18);
-//        }
-        Image icon = this.icon == null ?
-                (type == 0 ? null : theme.iconSet[type - 1]) : this.icon;
-        if (icon != null) {
-            rd.drawImage(icon, orientation == RIGHT_TO_LEFT ? (w - 7 - 32) : 6, (h / 2) - 15, 32, 32, null);
-        }
-        rd.setTransform(trans);
-        rd.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_OVER));
-    }
-
-    private int getLineHeight(Font font) {
-        FontMetrics ftm = DesktopNotifyDriver.getFontMetrics(font);
-        return ftm.getHeight() - ftm.getLeading();
-    }
-
-    /**
-     * Splits the words of the text across lines that fit inside the
-     * notification. It is called by the {@link DesktopNotifyDriver}.
-     */
-    protected void sortMessage() {
-        if (!title.isEmpty()) tlts = splitLines(title, theme.titleFont);
-        if (!message.isEmpty()) msgs = splitLines(message, theme.contentFont);
-        h = 15 + (getLineHeight(theme.titleFont) * tlts.length)
-                + (getLineHeight(theme.contentFont) * msgs.length);
-    }
-
-    private String[] splitLines(String in, Font font) {
-        String[] out;
-        ArrayList<String> list = new ArrayList<>();
-        String[] strs = in.split("\n");
-        StringBuilder builder = new StringBuilder();
-        FontMetrics ftm = DesktopNotifyDriver.getFontMetrics(font);
-        for (String str : strs) {
-            String[] words = str.split(" ");
-            for (String word : words) {
-                if (ftm.stringWidth(builder.toString()) + ftm.stringWidth(word)
-                        < (w - 12 - ((icon == null && type == 0) ? 0 : 38))) {
-                    builder.append(word).append(" ");
-                } else {
-                    list.add(builder.toString());
-                    builder = new StringBuilder().append(word).append(" ");
-                }
-            }
-            list.add(builder.toString());
-            builder.setLength(0);
-        }
-        out = new String[list.size()];
-        out = list.toArray(out);
-        return out;
-    }
-
-    // Integrated Logging
-    public static final int DEBUG = 0;
-    public static final int NONE = 10;
-
-    /**
-     * Allows to specify the level allowed for logging messages. Attempts to log
-     * messages with a priority level below what's been set will be ignored.
-     *
-     * @param level The logging priority level to allow as minimum.
-     */
-    public static void setLogLevel(int level) {
-        logger.setLoggingLevel(level);
-    }
-
-    /**
-     * Returns the level allowed for logging messages. Attempts to log
-     * messages with a priority level below what's been set will be ignored.
-     *
-     * @return The logging priority level curently allowed as minimum.
-     */
-    public static int getLogLevel() {
-        return logger.getLoggingLevel();
-    }
-
-    public static void setLogOutput(OutputStream outStream) {
-        logger.setOutput(outStream);
-    }
-
-    /**
-     * Logs a debug message. Parameters can't be {@code null}.
-     *
-     * @param tag     The tag associated to this message.
-     * @param message The message to include.
-     */
-    public static void logDebug(String tag, String message) {
-        logger.post(DEBUG, tag, message, null);
-    }
-
-    /**
-     * Logs an information message. Parameters can't be {@code null}.
-     *
-     * @param tag     The tag associated to this message.
-     * @param message The message to include.
-     */
-    public static void logInfo(String tag, String message) {
-        logger.post(INFORMATION, tag, message, null);
-    }
-
-    /**
-     * Logs a warning message. Parameters can't be {@code null}.
-     *
-     * @param tag     The tag associated to this message.
-     * @param message The message to include.
-     */
-    public static void logWarning(String tag, String message) {
-        logger.post(WARNING, tag, message, null);
-    }
-
-    /**
-     * Logs an error message. Parameters other than the caught throwable can't
-     * be {@code null}.
-     *
-     * @param tag     The tag associated to this message.
-     * @param message The message to include.
-     * @param tr      A Throable object (an Exception, for example), whose
-     *                message and stack trace should be logged too.
-     */
-    public static void logError(String tag, String message, Throwable tr) {
-        logger.post(ERROR, tag, message, tr);
-    }
-
-    private static final Logger logger = new Logger();
-
-    /**
-     * Utilitary subclass used for logging.
-     */
-    private static class Logger {
-        private OutputStream outStream;
-        private PrintWriter outWritter;
-
-        private int level;
-
-        Logger() {
-            this.level = INFORMATION;
-            this.setOutput(System.out);
-        }
-
-        /**
-         * Lets you specify an OutputStream for use in this Logger.
-         *
-         * @param outStream The OutputStream you wish to log messages into.
-         */
-        public synchronized void setOutput(OutputStream outStream) {
-            this.outStream = outStream;
-            this.outWritter = new PrintWriter(outStream);
-        }
-
-        /**
-         * @return <code>true</code> if a proper output destiny has been set, like
-         * an output file or stream.
-         */
-        public boolean isReady() {
-            return outWritter != null;
-        }
-
-        private String levelToString(int level) {
-            String str = "OTHER";
-            switch (level) {
-                case DEBUG:
-                    str = "DEBUG";
-                    break;
-                case INFORMATION:
-                    str = "INFO";
-                    break;
-                case WARNING:
-                    str = "WARNING";
-                    break;
-                case ERROR:
-                    str = "ERROR";
-                    break;
-            }
-            return str;
-        }
-
-        public synchronized void setLoggingLevel(int level) {
-            this.level = level;
-        }
-
-        public synchronized int getLoggingLevel() {
-            return this.level;
-        }
-
-        /**
-         * Checks for you if a given level is enabled in this logger.
-         *
-         * @param level The level to check.
-         * @return <code>true</code> if the given level is found to be enabled,
-         * <code>false</code> otherwise.
-         */
-        public boolean isEnabled(int level) {
-            return this.level <= level;
-        }
-
-        /**
-         * Full form of the method. This method lets you post a message into this
-         * logger, so it can be logged or not depending of the assigned settings.
-         * You can specify a title (tag), message, flag and a Throwable object (like
-         * an Exception or Error) to include in the log.
-         *
-         * @param level   The level to mark this message with.
-         * @param tag     The tag associated to this message. It cannot be null.
-         * @param message The message to include. It cannot be null.
-         * @param tr      A Throwable whose message and stack trace should be
-         *                printed.
-         */
-        public synchronized void post(int level, String tag, String message, Throwable tr) {
-            if (tag == null) throw new NullPointerException("Tag please");
-            if (message == null) throw new NullPointerException("Message please");
-            if (isReady() && isEnabled(level)) {
-                StringBuilder builder = new StringBuilder();
-                builder.append(levelToString(level)).append(":").append(tag)
-                        .append(" at [").append(new java.util.Date())
-                        .append("] -> ").append(message.trim());
-                if (outStream instanceof PrintStream) {
-                    ((PrintStream) outStream).println(builder);
-                } else {
-                    outWritter.println(builder);
-                    outWritter.flush();
-                }
-                if (tr != null) {
-                    if (outStream instanceof PrintStream) {
-                        ((PrintStream) outStream).println("Caused by:");
-                        tr.printStackTrace((PrintStream) outStream);
-                    } else {
-                        outWritter.println("Caused by:");
-                        tr.printStackTrace(outWritter);
-                        outWritter.flush();
-                    }
-                }
-            }
-        }
-    }
 }
