+++ src/main/java/ds/desktop/notify/Utils.java
@@ -142,7 +142,7 @@ public class Utils {
     public static boolean isTranslucencySupported() {
         boolean nativeTrans;
         if (System.getProperty("java.version").contains("1.6")) {
-            DesktopNotify.logWarning("NotifyUtils", "Per-pixel translucency is currently not "
+            DSLogger.logWarning("NotifyUtils", "Per-pixel translucency is currently not "
                     + "supported.\nPlease upgrade your JRE to at least Java 7 "
                     + "to support this feature.");
             nativeTrans = false;
@@ -152,7 +152,7 @@ public class Utils {
             GraphicsDevice gd = ge.getDefaultScreenDevice();
             if (!gd.isWindowTranslucencySupported(GraphicsDevice
                     .WindowTranslucency.PERPIXEL_TRANSLUCENT)) {
-                DesktopNotify.logError("NotifyUtils", "Error while starting: "
+                DSLogger.logError("NotifyUtils", "Error while starting: "
                         + "Per-pixel translucency is not supported.", null);
                 nativeTrans = false;
             } else nativeTrans = true;
